Adapt configuration object
