+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/internal/utils/AnnotationUtils.kt
@@ -8,6 +8,7 @@ import io.github.freya022.botcommands.api.commands.annotations.UserPermissions
 import io.github.freya022.botcommands.api.commands.application.annotations.Test
 import io.github.freya022.botcommands.api.core.BContext
 import io.github.freya022.botcommands.api.core.Filter
+import io.github.freya022.botcommands.api.core.config.applicationConfig
 import io.github.freya022.botcommands.api.core.service.getService
 import io.github.freya022.botcommands.api.core.utils.*
 import net.dv8tion.jda.api.Permission
